checkbox styling done
