+++ styles/globals.css
@@ -35,5 +35,3 @@
     @apply text-base leading-normal font-body;
   }
 }
-
-/* INPUT FIELDS */
